requirements.txt
langchain>=0.1.20
langchain-community>=0.0.38
langchain-openai>=0.1.7
langgraph>=0.0.65
chromadb>=0.4.24
python-dotenv>=1.0.1
streamlit>=1.33.0
openai>=1.30.1
requests>=2.32.2
docling>=2.63.0
pandas>=2.0.0
supabase>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.0.0